Moves GFF class to io.
