+++ scripts/gene_annotation.py
@@ -1,6 +1,7 @@
 import re
 
 from lcdblib.utils.chrom_convert import import_conversion
+from larval_gonad.io import GffRow
 
 GTF = snakemake.input[0]
 OUTNAME = snakemake.output[0]
@@ -31,27 +32,3 @@ def main():
                     )
                     + "\n"
                 )
-
-
-class GffRow(object):
-    def __init__(self, row):
-        self.seqid, self.source, self.type, self.start, self.end, self.score, self.strand, self.phase, self.attributes = row.strip().split(
-            "\t"
-        )
-        self.is_gene = self.type == "gene"
-        self.parsed_attributes = self.parse_attributes()
-
-    def parse_attributes(self):
-        parsed_attributes = {}
-        for attr in self.attributes.split(";"):
-            mm = re.search('(?P<key>.*?)\s+"(?P<value>.*?)"', attr)
-            if mm:
-                parsed_attributes[mm.group("key").strip()] = mm.group("value").strip()
-        return parsed_attributes
-
-    def __getitem__(self, key):
-        return self.parsed_attributes[key]
-
-
-if __name__ == "__main__":
-    main()
